refactor: split tool handlers into pipeline/data/interactive modules (Task 19)
Decompose tools_v2.py (1276 lines) into focused submodules:
- tools/data.py: data source and platform handlers
- tools/pipeline.py: batch pipeline handlers
- tools/interactive.py: interactive shipment handler
- tools/__init__.py: canonical tool registry with get_all_tool_definitions()
- tools_v2.py: temporary barrel re-exporting all names (removal target: 2026-03-14)

All test patch targets updated to reference new submodule locations.
945 passed, 5 skipped (integration/mcp errors are pre-existing).
